scrapper/binance_Klines_producer/requirements.txt
kafka-python-ng==2.2.3
requests==2.32.3
urllib3==2.2.3